requirements.txt
pip~=23.3.1
wheel~=0.41.3
pillow~=11.1.0
docutils~=0.19
jinja2~=3.1.6
opencv-python~=4.10.0.84
numpy~=1.26.2
pyserial~=3.5
ultralytics~=8.3.134